Brings in main's four post-CUDA PRs: the genus-95 curve code (#27), the
AG-skip zerocheck wired end to end (#28/#31), the all-core (P+E) pool with
topology-gated hops (#29), and the Ligerito initial-sumcheck lookahead +
NTT deep-pass package (#30).

Resolution notes, beyond the mechanical hunks:

- NTT deep pass: main's fused 2-layer restructure adopted with this
  branch's live-lane threading; fused_2layer_row_op now iterates the row
  slice's own length (the textual auto-merge would have indexed dead lanes
  out of bounds when live < num_ntts).
- Ligerito lookahead: machinery + tests kept; the F128 driver main
  scheduled it in was already replaced here by the F256-ladder driver
  (extension.rs), which accepts and ignores the coefficients — transcript-
  neutral by exact polynomial identity. Porting the two-rounds-per-pass
  schedule onto the ladder's fused folds is the recorded follow-up.
- SkipPoint: threaded through the lincheck grinding (the fused PoW+squeeze
  transcript is unchanged on the phi8 arm); phi8-only protocol surfaces
  (merged transport, union lincheck, union claim-point builders) convert
  at the boundary via .phi8().
- The retired chain product stays deleted; main's small edits to those
  files (SkipPoint adaptations) die with them. Blake3Setup regains the
  generate_witness_ab dispatch and the prove_fast_timed wrapper that the
  AG methods and the blake3_rs_vs_ag bench consume.
- AG open calls route through the grinding-aware open with
  pcs_params.opening_grinding() (disabled on default profiles, so main's
  AG transcripts are unchanged).
- Workspace-wide clippy cleanup so `-D warnings` stays green on current
  stable.

Verified: cargo fmt --check and clippy --release -D warnings clean;
cargo test --release --workspace green (674 passed). Byte pins: the m6
merged-union fixtures pass; the four ignored failures (transcript_shape's
element pin, union_element's three mixed tests) fail identically on
pre-merge HEAD with identical computed hashes — pre-existing fixture rot,
not moved bytes.

+++ crates/flock-core/src/element_r1cs.rs
@@ -716,7 +716,7 @@ impl Grinding {
     /// `m_words / (2^bits |F|) < 2^-128` for `m_words >= 1`.
     pub fn initial_bits(self, m_words: usize) -> Option<u32> {
         self.enabled
-            .then_some((usize::BITS - m_words.leading_zeros()) as u32)
+            .then_some(usize::BITS - m_words.leading_zeros())
     }
 
     pub fn round_bits(self) -> Option<u32> {
